This is your fucking commit

+++ Core/Inc/encoder.hpp
@@ -8,7 +8,7 @@ public:
   int16_t cpr; // count per revolution
   uint32_t sample_time; // time to sample for omega
   float omega;
-  int64_t count_aggregate = 0;
+  int32_t count_aggregate = 0;
   uint32_t last_reset_time = 0;
   int dir;
 
@@ -26,6 +26,6 @@ public:
   void init();
   float get_encoder_omega();
   int32_t get_count();
-  int64_t get_count_aggregate();
+  int32_t get_count_aggregate();
   void reset_encoder_count();
 };
